Refactoring: Moved classes for reporting into their own folder. This is to make locating and understanding project structure easier.
