+++ server/routes/reporting/performanceReport/reportingPresenter.ts
@@ -1,5 +1,5 @@
-import PresenterUtils from '../../utils/presenterUtils'
-import { FormValidationError } from '../../utils/formValidationError'
+import PresenterUtils from '../../../utils/presenterUtils'
+import { FormValidationError } from '../../../utils/formValidationError'
 
 export default class ReportingPresenter {
   constructor(
