+++ server/routes/serviceProviderReferrals/serviceProviderReferralsController.ts
@@ -73,10 +73,10 @@ import InitialAssessmentFeedbackConfirmationPresenter from '../appointments/feed
 import InitialAssessmentFeedbackConfirmationView from '../appointments/feedback/initialAssessment/confirmation/initialAssessmentFeedbackConfirmationView'
 import ActionPlanSessionBehaviourFeedbackPresenter from '../appointments/feedback/actionPlanSessions/behaviour/actionPlanSessionBehaviourFeedbackPresenter'
 import InitialAssessmentBehaviourFeedbackPresenter from '../appointments/feedback/initialAssessment/behaviour/initialAssessmentBehaviourFeedbackPresenter'
-import ReportingPresenter from './reportingPresenter'
-import ReportingView from './reportingView'
-import ReportingForm from './reportingForm'
-import PerformanceReportConfirmationView from './performanceReportConfirmationView'
+import ReportingPresenter from '../reporting/performanceReport/reportingPresenter'
+import ReportingView from '../reporting/performanceReport/reportingView'
+import ReportingForm from '../reporting/performanceReport/reportingForm'
+import PerformanceReportConfirmationView from '../reporting/performanceReport/confirmation/performanceReportConfirmationView'
 import DraftsService, { Draft } from '../../services/draftsService'
 import AuthUserDetails from '../../models/hmppsAuth/authUserDetails'
 import { ActionPlanAppointment, AppointmentSchedulingDetails } from '../../models/appointment'
